Updates README to reference new location
This is the only link I find when searching for elm-test in Google. It's very difficult for other people to see/understand that this package has been moved to the elm-community. I'm not sure if this also needs to exists, but I'v created a "deprecated" text with a link to the new location. If it is not possible to have the DEPRECATED text, could you please provide a link in the readme to the updated solution?